[libc] Add support to compile some syscalls on 32 bit platform
Automated commit created by applying diff 546523

Phabricator-ID: PHID-HMBT-gledxeq2dmpen4qoyatc
Review-ID: D148371

+++ libc/test/src/sched/sched_rr_get_interval_test.cpp
@@ -29,7 +29,8 @@ TEST(LlvmLibcSchedRRGetIntervalTest, SmokeTest) {
   };
 
   auto TimespecToNs = [](struct timespec t) {
-    return t.tv_sec * 1000UL * 1000UL * 1000UL + t.tv_nsec;
+    return static_cast<unsigned long>(t.tv_sec * 1000UL * 1000UL * 1000UL +
+                                      t.tv_nsec);
   };
 
   struct timespec ts;
